@@ -44,7 +44,7 @@
         }
     },
     "require": {
-        "php": ">=7.2",
+        "php": ">=7.4",
         "ext-json": "*",
         "psr/log": "^1.1 || ^2.0 || ^3.0",
         "symfony/deprecation-contracts": "^2 || ^3",
@@ -62,6 +62,7 @@
         "friendsofphp/php-cs-fixer": "^2.17 || ^3.47.1",
         "phpstan/phpstan": "^1.6",
         "phpunit/phpunit": ">=8",
+        "rector/rector": "^1.0",
         "vimeo/psalm": "^4.23"
     },
     "suggest": {
